fix(spawn): keep route exports Next compatible
